refactor quantization and metrics of INC
