+++ python/nano/src/bigdl/nano/quantization/neural_compressor/__init__.py
@@ -13,3 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+from .quantization import QuantizationINC
