+++ python/nano/src/bigdl/nano/quantization/neural_compressor/metric/__init__.py
@@ -13,3 +13,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+
+from .metrics import *
+
+METRICS = {
+    'pytorch': PytorchINCMetric,
+    'tensorflow': TensorflowINCMetric
+}
